Swapped some bits around. FIxed reloading
